requirements.txt
pygame>=2.5.0
PyOpenGL>=3.1.6
numpy>=1.23.0
pyyaml>=6.0